Add UI to delete scan files

+++ frontend/distiller/src/components/location.tsx
@@ -7,8 +7,8 @@ import { ScanLocation, IdType } from '../types';
 
 import {  } from '../features/scans';
 
-import { useAppDispatch } from '../app/hooks';
-import { removeScanFiles } from '../features/scans';
+import { useAppDispatch, useAppSelector } from '../app/hooks';
+import { removeScanFiles, scanSelector } from '../features/scans';
 
 import { COMPUTE_HOSTS } from '../constants';
 
@@ -19,7 +19,7 @@ const useStyles = makeStyles((_theme) => ({
 type Props = {
     scanID: IdType;
     locations: ScanLocation[];
-
+    confirmRemoval: (scanID: number, title: string, message: string) => Promise<boolean>;
 }
 
 type UniqueLocation = {
@@ -28,20 +28,28 @@ type UniqueLocation = {
 }
 
 type ChipProps = {
-  scanID: number,
+  scanID: number;
   host: string;
+  confirmRemoval: (scanID: number, title: string, message: string) => Promise<boolean>;
 }
 
-
 const LocationChip: React.FC<ChipProps> = (props) => {
-
   const dispatch = useAppDispatch();
   const [deletable, setDeletable] = React.useState(true);
-  const {scanID, host} = props;
+  const {scanID, host, confirmRemoval} = props;
+  const scan = useAppSelector(scanSelector(scanID));
 
-  const onDelete = () => {
-    dispatch(removeScanFiles({id: scanID, host}));
-    setDeletable(false);
+  const onDelete =  async () => {
+    if (scan === undefined) {
+      return;
+    }
+
+    const confirmed = await confirmRemoval(scanID, "Remove scan files", `You are about to remove scan files from the acquisition machine for scan ${scan.scan_id}. This operation can not be undone.`);
+
+    if (confirmed) {
+      dispatch(removeScanFiles({id: scanID, host}));
+      setDeletable(false);
+    }
   };
 
   return (
@@ -65,13 +73,12 @@ const LocationComponent: React.FC<Props> = (props) => {
     return locs;
   }, {} as {[host: string]: UniqueLocation}));
 
-
   return (
     <React.Fragment>
       {uniqueLocations.map(location => {
         return (
           <div key={location.host} title={location.paths.join(', ')} className={classes.chip}>
-            <LocationChip scanID={props.scanID} host={location.host}/>
+            <LocationChip scanID={props.scanID} host={location.host} confirmRemoval={props.confirmRemoval}/>
           </div>
         )
       })}
